Perk by Buglight
Still not sure why you're reading this. Are you some kind of FBI agent? Anyways, R.I.P. the 4th anniversary event and all the gateaus I will burn for the next month. Last version had some slight bugs, and unfortunately I wasn't quite sure how to reproduce them, but I believe I fixed them in the newest version. Most of everything to do with icon packs are fully functional *except* updating them afterwards, which will be coming soon. An explanation as to why that's not a thing yet:

I'd like updating an icon pack with new images to be about as painless as possible and in order to do that, one must take said image *in a custom dialog box, mind you,* and provide categories such as icon type (Power, Portrait, Addon, Status Effect, etc.), and any sub-categories such as character. After that, I'll *probably* check the image width and height to make sure you're not trying to put a 4K image where a 256x256 image used to be. Probably not necessary since you could do it manually, but at least there will be some checks and balances. Anyways, on to the patch notes!

This will be marked as PbD v0.1.1
The previous version, PbD v0.1.0 did not get a release since there were bugs that made the update somewhat dysfunctional.
I'm planning to release this version, not marked as pre-release (or beta).

* pom.xml was updated to remove Google Guava v19.0 as a dependency, which posed a minor security issue *if, and only if* there were certain classes used (which I did not) and if the dependency was hosted on a server (which it is not).
* README was updated to current standards of the project.
+ Power is a sub-category of icon and will be displayed under the icons list.
+ StatusEffect is a sub-category of icon and will be displayed under the icons list.
- Files has been (re)moved to the utils package.
- ZipEvaluator has been removed entirely.
* IconPackTreeView has been updated for the new icon categories Power and StatusEffect.
* IconInstallTask was updated to debug the paths which was related to the previous version and will be removed later on.
* MoveTask forgot to delete the main folder that it was supposed to copy.
* UnzipTask was updated to make sure the parent file existed before actually extracting the file.
* ErrorLog was updated to get rid of a lil typo.
* PBD
  - Removed (now) useless code that generated an enum for me because I'm lazy.
  - Added Power and StatusEffect to the set of icons.
* Files was updated for further checks on a copy.