v0.2.2 - Documentation & Download Update
Hello all, and a shout out to DBD for giving me 250,000 blood points for an error almost no one came across but still haven't given me pips back for the game crashing. /rant over

This update is primarily focused on heavily documenting Perk by Daylight as well as making updates a hell of a lot nicer for the user.

This version of Perk by Daylight works with icon packs that are up-to-date with Dead by Daylight patch 7.1.3. An update to support the new survivor and killer will be made in the near future, but since those characters are still in the PTB, I've made this update a priority over that.

* Most classes, if not listed below, have only been modified by adding documentation for classes/methods/parameters.
* MainController now automatically checks for an update if updates are not ignored and if the user is not in offline mode. If either of those, or both, are checked, the program will not check for an update.
* Merged CopyTask and MoveTask into a single class and added a "delete" param in the constructor to define if the original files should be deleted after being copied.
+ Added UpdateForm and UpdateController. This is a form that will be presented when UpdateTask returns true, meaning there is an update available.
+ Added DownloadTask, the task used to download the update. This task will automatically delete the old version and replace it with a new one a couple seconds after closing. Make sure not to mess with anything until this is completed.
* Latest is now deprecated due to the "new" format of a latest.json file hosted on glitchedcode.dev no longer being applicable due to the lack of a big enough audience. Sorry lads, but it seems not many people are actually using this program.